requirements.txt
pymodbus==2.5.3
Pillow==9.5.0
RPi.GPIO==0.7.1
